double promotion warning fix or ignore per module

+++ src/modules/events/send_event.cpp
@@ -263,8 +263,8 @@ int SendEvent::custom_command(int argc, char *argv[])
 
 		struct vehicle_command_s cmd = {
 			.timestamp = 0,
-			.param5 = (float)((accel_calib || calib_all) ? vehicle_command_s::PREFLIGHT_CALIBRATION_TEMPERATURE_CALIBRATION : NAN),
-			.param6 = NAN,
+			.param5 = ((accel_calib || calib_all) ? vehicle_command_s::PREFLIGHT_CALIBRATION_TEMPERATURE_CALIBRATION : (double)NAN),
+			.param6 = (double)NAN,
 			.param1 = (float)((gyro_calib || calib_all) ? vehicle_command_s::PREFLIGHT_CALIBRATION_TEMPERATURE_CALIBRATION : NAN),
 			.param2 = NAN,
 			.param3 = NAN,
