+++ src/modules/navigator/mission_feasibility_checker.cpp
@@ -522,8 +522,8 @@ MissionFeasibilityChecker::checkDistancesBetweenWaypoints(const mission_s &missi
 		return true;
 	}
 
-	double last_lat = NAN;
-	double last_lon = NAN;
+	double last_lat = (double)NAN;
+	double last_lon = (double)NAN;
 
 	/* Go through all waypoints */
 	for (size_t i = 0; i < mission.count; i++) {
